Added webview .py & .sh, former in python(3), latter in posix shell
What these do: given argument, replace call with one to image viewer or browser
depending on whether the input is an image or not.

Both should do the same thing, the python one was made first though. Although it
is strange that the python one did _not_ wowk with palemoon despite having the
correct path.

webivew was made in mind to RTV which was said to rely on $RTV_BROWSER (didn't
work for me?). Didn't want a giant browser to open when all I was viewing was an
image. Ideally one could replicate it to do the same with video files, etc.

On the other hand, perhaps a test of prefix, to open certain domains differently.
E.g. like route youtube links to youtube-dl or alternatively my fq script (go).