fq.pl now working with 'add' and 'view' cmds. Which 'add' urls to the dbm queue and 'view' downloaded videos. Only youtube-dl is supported ATM with hard coded format codes. (a basic web search shows reading env-vars to be possible)
An addition is that fq.pl can manage more than 2 videos in its 'cache',
as opposed to fq9 & fq.py which can only do 2. This is done by piggy
backing on dqueue's dbm file to hold a filename to write to next.
More testing is needed (which is completed after I watch 6+ videos
without clearing the files).

Next thing I want to do is output youtube-dl/other program output to
a log file rather than stdout. Output in this perl version is not as
good as the others so I'm going to seperate it.