Working prototype, call like so:
python3 yt_dl_urls URL HEIGHT [WEBM|MP4] ABR

ATM it saves both video and audio to
/mnt/ramdisk/1 and 1aud
and will fail if they already exists or the directory
does not exists.

What this is is a script that retrieves URLs from youtube
using youtube-dl, and then downloads the video and audio
with threads and then plays the result. For now I only
know that the downloading works for sure.

Script needs much work