Carried new knowledge to fq_t.py This being: 	the use of the format specifiers, 	rate limit, 	and threading.
I do not know why I have to use threading now when I did not before, but
I'll check the diff after this commit.
One major change getting rid of redundant functions that more or
less did the same thing (due to them using the same arguments)

The other would be getting rid of pointless returns (return None?).

A remaining problem is that the download occurs even if either doesn't exists.
Maybe I could use a try/except and end the other downloading thread in case this
happens?

yt_dl_urls.py (which is becoming less and less relevant) can do this but I'll
have to change the program a bit.