Added streamlink option, the only thing 'essential' fq10 is missing is the 'halt' cmd, which needs improvement
Changed 'url' from global to local variable.

Notes:
apparantly there are costs with using subshell like I do, case in point,
	last=$(getHistUrl)
the 'fix' being to use a global var to hold the 'return' value
but, I'm not going to do it, after all the whole reason I started using
locals in fq10 (and not before) was because it reduced the confusion of
when/where a particular variable was set. (tryResume is an exception)
I suppose I /could/ use a return value to say that func(x) was successful in
setting the var, but that seems to be a lot of trouble for little to no
gain for fq's case

Instead of setting cmd to 'dl' when a download is set, tryDl is now set to
true. (also a global var) 'cmd' should not have been reused like that to
begin with.