Attempted to add resume support to streamlink, the code is right, but the stream object is not seekable meaning resume cannot be implemented
Extracted the 'url recognizer' and reworked it to be more expandable.

Made an observation that serStreamlink/youtube (or what used to be case)
is only called by the autoServicer meaning it does not need to check for
valid url or take arguments like the other case functions do, planning to
refactor it next.