Made a download function that uses tandem_download, to do this I had to make functions that got the url instead of the format codes, the previous ones were taken and modified to return the whole object (the legacy funcs call the new funcs)
A note about tandem download is that all downloads will go at the same rate meaning it
is very possible that the audio will finish way before the video due to filesize differences.

In the past I differentiate the speeds to remedy this, but tandem download doesn't allow this
at the moment (or even easily)

A hackish way of doing it may be to give a ratio and add the same ChunkStreamDownloader to the
download list multiple times to enforce this ratio. (e.g. [vid, vid, vid, aud] would download
the video 3x faster). This is me guessing, of course.
What will this do to the actual ratelimit is anyone's guess.

Todo: give tandem_download's caller better names. I don't plan on discarding download_by_thread