Moved out streamlink download-limiter to standalone thing (that still depends on streamlink itself)
Has basic argparser for its basic arguments, which are all required.
Formats/qualities/resolutions must be gotten via streamlink itself as
this does not have the capability at the moment.

In the future, fq_t AND fq10 will depend on streamlink_dl since the
former already does in way, and the latter needs download rate limiting.